requirements.txt
fastmcp>=0.4.0
langgraph>=0.2.0
langchain-ollama>=0.2.0
langchain-mcp-adapters>=0.1.0
langchain-core>=0.3.0
tavily-python>=0.5.0
pydantic>=2.0.0
